added Division

+++ calculations.cpp
@@ -5,23 +5,23 @@ Calculations::Calculations()
 
 }
 
-int Calculations::addition(int currentNumber, int numberToBeAdded)
+double Calculations::addition(double currentNumber, double numberToBeAdded)
 {
     return currentNumber + numberToBeAdded;
 
 }
 
-int Calculations::subtraction(int currentNumber, int numberToBeAdded)
+double Calculations::subtraction(double currentNumber, double numberToBeAdded)
 {
     return currentNumber - numberToBeAdded;
 }
 
-int Calculations::division(int currentNumber, int numberToBeAdded)
+double Calculations::division(double currentNumber, double numberToBeAdded)
 {
     return currentNumber / numberToBeAdded;
 }
 
-int Calculations::multiplikation(int currentNumber, int numberToBeAdded)
+double Calculations::multiplikation(double currentNumber, double numberToBeAdded)
 {
     return currentNumber * numberToBeAdded;
 }
